Added Requester class
Added Requester class and moved all request related logic from RavenClient to it. Also added a RavenClient.BeforeSend func so the Requester is exposed and can be manipulated before the HTTP request is performed.

Also added RequestData class to capture the HTTP body in scrubbed and raw variants to simplify debugging and logging.