Demangle Function Name from Async Calls
When the Microsoft cs.exe compiler compiles some modern C# features,
such as async/await calls, it can create synthetic function names that
do not match the function names in the original source code. Here we
reverse some of these transformations, so that the function and module
names that appears in the Sentry UI will match the function and module
names in the original source-code.

Fixes #147