Add support for submitting User Feedback to Sentry
This commit adds support for submitting user feedback to Sentry.
The feedback form post API endpoint is used, which is undocumented, so
it is possible this may break. However the endpoint has been stable for
the past 7 months or so.

Notes:

* The sentry form post API endpoint is found at https://<sentry url>/api/embed/error-page/?dsn=<dsn>&eventId=<eventId>name=<name>&email=<email>&comments=<comments>
* All form fields must be filled out (name, email, comments), and each user feedback must have
an associated eventId

Changes:

Add the SentryUserFeedback class which holds the feedback information
(Name, Email, Comment, EventID)

Modify the Requester class:

* Add a user feedback specific constructor, which takes a
SentryUserFeedback

* Add the CreateWebRequest method, which takes a URI and returns an
HttpWebRequest object with the boilerplate information setup (timeout,
authorization etc.). This reduces copy-and-pasting code between the two
constructors

* Add the SendFeedback method, which sends the user feedback to Sentry

Modify the Requester.Net45 class:

* Add the SendFeedbackAsync method, which sends the user feedback to
Sentry asynchronously

Modify the RavenClient class:

* Add the SendUserFeedback method, which takes a SentryUserFeedback,
creates the Requester object for the feedback, and sends it to Sentry

Modify the RavenClient.Net45 class:

* Add the SendUserFeedbackAsync method, which does the same job as the
SendUserFeedback method, but asynchronously

Modify the Dsn class:

* Add the FeedbackUri property, which is the user feedback form
submission endpoint